Fixed index out of range iteration error on Delete Tasks; Added formatting checks on property inputs; changed Block & Type properties to Minutes & Importance;  Addd table drawing output of task list

+++ TaskMate/SpectreUI.cs
@@ -24,7 +24,7 @@ namespace TaskMate
                     AddTaskMenu();
                     break;
                 case "Display Tasks":
-                    DisplayTasks();
+                    DisplayTasksTable();
                     break;
                 case "Edit Tasks":
                     SelectEditTask();
@@ -42,19 +42,50 @@ namespace TaskMate
 
         public static void AddTaskMenu()
         {
+            bool intFlag = false;
+            int minutes =  0;
+            int importance = 0;
+
             Console.Clear();
             AnsiConsole.WriteLine("Task Name?");
             String name = Console.ReadLine();
-            AnsiConsole.WriteLine("Task Time Blocks?");
-            String blocks = Console.ReadLine();
-            AnsiConsole.WriteLine("Task Type?");
-            String type = Console.ReadLine();
+            
+            while (intFlag == false)
+            {
+                AnsiConsole.WriteLine("How many minutes will this take?");
+                String min = Console.ReadLine();
+                if (!int.TryParse(min, out minutes))
+                {
+                    Console.WriteLine("Please enter a valid number.");                    
+                }
+                else intFlag = true;
+            }
+
+
+
+            intFlag = false;
+            while (intFlag == false)
+            {
+                AnsiConsole.WriteLine("Level of importance (1-4)?");
+                String imp = Console.ReadLine();
+                if (!int.TryParse(imp, out importance))
+                {
+                    Console.WriteLine("Please enter a valid number.");
+                    intFlag = false;
+                }
+                else if (!(importance > 1 && importance < 4))
+                {
+                    Console.WriteLine("Please enter a valid number.");
+                    intFlag = false;
+                }
+                else intFlag = true;
+            }
 
             var task = new Task()
             {
                 Name = name,
-                Blocks = blocks,
-                Type = type
+                Minutes = minutes,
+                Importance = importance
             };
 
             Repo.AddTask(task); ;
@@ -65,10 +96,28 @@ namespace TaskMate
             foreach (var task in Repo.Load()) PrintTask(task);
             static void PrintTask(Task task)
             {
-                AnsiConsole.WriteLine($"Name: {task.Name}\t Blocks: {task.Blocks}\t Type: {task.Type}");
+                AnsiConsole.WriteLine($"Name: {task.Name}\t Minutes: {task.Minutes}\t Type: {task.Importance}");
             }
         }
 
+        public static void DisplayTasksTable()
+        {
+            var table = new Table();
+            table.Border = TableBorder.Simple;
+
+            table.AddColumn("Name");
+            table.AddColumn("Minutes");
+            table.AddColumn("Type");
+                      
+            foreach (Task t in Repo.Load())
+            {
+                var name = t.Name;
+                var minutes = t.Minutes.ToString();
+                var importance = t.Importance.ToString();
+                table.AddRow(name, minutes, importance);
+            }           
+            AnsiConsole.Write(table);
+        }
 
         public static void SelectEditTask()
         {
@@ -88,24 +137,27 @@ namespace TaskMate
                     .AddChoices(new[] {"", "<- BACK"}));
             //.MoreChoicesText("[grey](Move up and down to reveal more tasks)[/]")
 
-            var taskList = Repo.Load();
-            int found = selection.IndexOf(".");
-            int length = selection.Length;
-            int id = int.Parse(selection.Remove(found, length - found));
-            
-            SelectEditProperty(id);
+            if (selection != "<- BACK" && selection != "")
+            {
+                var taskList = Repo.Load();
+                int found = selection.IndexOf(".");
+                int length = selection.Length;
+                int id = int.Parse(selection.Remove(found, length - found));
 
+                SelectEditProperty(id);
+            }
         }
 
         public static void SelectEditProperty(int id)
         {
             Task task = Repo.Load()[id];
+            string s = string.Empty;
             var selection = AnsiConsole.Prompt(
                 new SelectionPrompt<string>()
                     .Title("What do you want to edit?")
                     .PageSize(10)
                     .MoreChoicesText("[grey](Move up and down to reveal more fruits)[/]")
-                    .AddChoices(new[] { task.Name, task.Blocks, task.Type, "<- Back", })
+                    .AddChoices(new[] { task.Name, task.Minutes.ToString(), task.Importance.ToString(), "<- Back", })
                 );
             
             if (selection == "<- Back") return;
@@ -114,17 +166,20 @@ namespace TaskMate
                 task.Name = AnsiConsole.Ask<string>("Task [green]name[/]?");
                 Repo.UpdateTask(task, id);
             }
-            else if (selection == task.Blocks)
+            else if (selection == task.Minutes.ToString())
             {
-                task.Blocks = AnsiConsole.Ask<string>("Task [green]blocks[/]?");
+                s = AnsiConsole.Ask<string>("Task [green]minutes[/]?");
+                task.Minutes = int.Parse(s);
                 Repo.UpdateTask(task, id);
             }
-            else if (selection == task.Name)
+            else if (selection == task.Importance.ToString())
             {
-                task.Type = AnsiConsole.Ask<string>("Task [green]type[/]?");
+                s = AnsiConsole.Ask<string>("Task [green]importance[/]?");
+                task.Importance = int.Parse(s);
                 Repo.UpdateTask(task, id);
             }          
         }
+
         public static void SelectDeleteTask()
         {
             List<string> nameList = new List<string>();
@@ -149,14 +204,30 @@ namespace TaskMate
             //int id = int.Parse(selection.Remove(found, length - found));
 
             //SelectEditProperty(id);
-
-            foreach (var item in selectionList)
+            //var selectionListR = selectionList.Reverse;
+            for (int i = selectionList.Count - 1; i >= 0; i--)
             {
-                int found = item.IndexOf(".");
-                int length = item.Length;
-                int id = int.Parse(item.Remove(found, length - found));
-                Repo.DeleteTask(id);
+                int found = selectionList[i].IndexOf(".");
+                if (found != -1)
+                {
+                    int length = selectionList[i].Length;
+                    int id = int.Parse(selectionList[i].Remove(found, length - found));
+                    Repo.DeleteTask(id);
+                }
+
+
             }
+            //foreach (string item in selectionList)
+            //{
+            //    int found = item.IndexOf(".");
+            //    if (found != -1)
+            //        {
+            //        int length = item.Length;
+            //        int id = int.Parse(item.Remove(found, length - found));
+            //        Repo.DeleteTask(id);
+            //        }
+            //}
+
         }
     }
 }
